@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Mail } from '../entities/mail';
 import { User } from '../entities/user';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { UserService } from '../services/user.service';
 
 @Component({
