@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import { of } from 'rxjs/observable/of';
+import { Observable,  of } from 'rxjs';
 
 import { Data } from '../entities/data';
 import { DATA } from '../mocks/data-mock';
